Refactor: 커스텀 훅 수정 - IntersectingObserver 관련 유사한 코드 합침
