+++ .eslintrc.json
@@ -1,10 +1,8 @@
 {
-    "extends": [
-        "next/core-web-vitals",
-        "prettier"
-    ],
+    "extends": ["next/core-web-vitals", "prettier"],
     "parserOptions": {
-        "project": "./tsconfig.json"
+        "project": ["tsconfig.json"],
+        "sourceType": "module"
     },
     "rules": {
         "react/react-in-jsx-scope": "off"
